#3372 Check the status of the datasource that failed to ingest
